more fixes to types, as well as switching gcc tests to use -m32 to match our semantics' implementation-defined settings, which actually causes our version of gcc to fail on at least one torture test.  also, switched size_t to match -m32

+++ semantics/includes/stdio.h
@@ -1,5 +1,5 @@
 #define NULL ((void *)0)
-typedef unsigned long int size_t; // this needs to correspond to cfg:sizeut
+typedef unsigned int size_t; // this needs to correspond to cfg:sizeut
 
 typedef struct {
 	unsigned long long int offset;
