+++ semantics/includes/stdlib.h
@@ -2,7 +2,7 @@
 #define EXIT_FAILURE 1
 #define RAND_MAX 2147483647
 #define NULL ((void *)0)
-typedef unsigned long int size_t; // this needs to correspond to cfg:sizeut
+typedef unsigned int size_t; // this needs to correspond to cfg:sizeut
 typedef struct {int quot; int rem;} div_t;
 typedef struct {long int quot; long int rem;} ldiv_t;
